- cleanup x3 and refactored AppDelegate to remove unwanted widgetkit method
