+++ Client/Frontend/Browser/TopSitesManager.swift
@@ -7,6 +7,7 @@ import Shared
 import UIKit
 import Storage
 import SyncTelemetry
+import WidgetKit
 
 struct TopSitesHandler {
     static func getTopSites(profile: Profile) -> Deferred<[Site]> {      
@@ -46,6 +47,26 @@ struct TopSitesHandler {
             deferred.fill(newSites)
             
             return deferred
+        }
+    }
+    
+    @available(iOS 14.0, *)
+    static func writeWidgetKitTopSites(profile: Profile) {
+        TopSitesHandler.getTopSites(profile: profile).uponQueue(.main) { result in
+            var widgetkitTopSites = [WidgetKitTopSiteModel]()
+            result.forEach { site in
+                // Favicon icon url
+                let iconUrl = site.icon?.url ?? ""
+                let webUrl = URL(string: site.url)
+                let imageKey = site.tileURL.baseDomain ?? ""
+                widgetkitTopSites.append(WidgetKitTopSiteModel(title: site.title, faviconUrl: iconUrl, url: webUrl ?? URL(string: "")!, imageKey: imageKey))
+                // fetch favicons and cache them in disk
+                FaviconFetcher.downloadFaviconAndCache(imageURL: !iconUrl.isEmpty ? URL(string: iconUrl) : nil, imageKey: imageKey )
+            }
+            // save top sites for widgetkit use
+            WidgetKitTopSiteModel.save(widgetKitTopSites: widgetkitTopSites)
+            // Update widget timeline
+            WidgetCenter.shared.reloadAllTimelines()
         }
     }
